Lift regions along with `let`s
This transforms expressions of the form

```
let x =
  region (
    let y = E1 in
    E2
  )
in
E3
```

into

```
region (
  let y = E1 in
  let x = E2 in
  tail (
    E3
  )
)
```

so that `y` is available when simplifying `E3` (as is the advantage of lifting
`let`s to begin with). This requires that `E3` not already have a `tail`
expression (or tail call) inside an `if` branch or any other construct besides
the body of a `let`, `let mutable`, `region`, or `tail`.