Don't try to be clever with `let x = M in tail x`
There's nothing actually wrong with `let x = M in tail x`, so don't try to
reduce it to just `M`. This works in that narrow case, but the code that was
doing this transformation didn't notice if the body of the tail is more than
just `x`. Since the transformation doesn't actually gain anything, better to
be rid of it than to make things more complicated trying to get it right.

Also updated a few comments and added an invariant check.